Took Repo name from json file
