+++ storage/validateArtifactLowerCase/validateArtifactLowerCase.json
@@ -1,0 +1,1 @@
+["repositoryName1", "repositoryName2"]
